Move app restart(ed) to app actions
Because we're making a move to put all these things in the actions
instead of the API as it makes writing requests that rely on one-
another easier to write. This is getting the app start actions code
ready.

See discussion
https://github.com/18F/cg-dashboard/pull/944#pullrequestreview-15829786
for more information on this.
