+++ static_src/util/cf_api.js
@@ -278,7 +278,7 @@ export default {
 
   postAppRestart(appGuid) {
     return http.post(`${APIV}/apps/${appGuid}/restage`).then(() => {
-      appActions.restarted(appGuid);
+      return appGuid;
     }).catch((err) => handleError(err,
       appActions.error.bind(this, appGuid)));
   },
